--- src/main/java/com/pick/zick/config/JpaConfig.java
@@ -1,4 +1,0 @@
-package com.pick.zick.config;
-
-public class JpaConfig {
-}
