--- src/main/java/com/pick/zick/config/userdetails/SecurityUserDetailsService.java
@@ -1,20 +1,0 @@
-package com.pick.zick.config.userdetails;
-
-import com.pick.zick.domain.entity.User;
-import com.pick.zick.domain.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.*;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class SecurityUserDetailsService implements UserDetailsService {
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User u = userRepository.findById(username)
-                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다: " + username));
-        return SecurityUserDetails.from(u);
-    }
-}
