+++ src/main/java/com/pick/zick/domain/auth/controller/AuthController.java
@@ -1,0 +1,22 @@
+package com.pick.zick.domain.auth.controller;
+
+import com.pick.zick.domain.auth.dto.request.LoginRequest;
+import com.pick.zick.domain.auth.dto.response.LoginResponse;
+import com.pick.zick.domain.auth.service.AuthService;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/auth")
+public class AuthController {
+
+    private final AuthService authService;
+
+    @PostMapping("/login")
+    public ResponseEntity<LoginResponse> login(@RequestBody @Valid LoginRequest req) {
+        return ResponseEntity.ok(authService.login(req));
+    }
+}
