+++ src/main/java/com/pick/zick/domain/auth/dto/response/SignupResponse.java
@@ -1,0 +1,7 @@
+package com.pick.zick.domain.auth.dto.response;
+
+public record SignupResponse(String userId, String message) {
+    public static SignupResponse ok(String id){
+        return new SignupResponse(id, "회원가입이 완료되었습니다.");
+    }
+}
