+++ src/main/java/com/pick/zick/domain/user/dto/request/UpdateUserRequest.java
@@ -1,0 +1,10 @@
+package com.pick.zick.domain.user.dto.request;
+
+import jakarta.validation.constraints.Size;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.Max;
+
+public record UpdateUserRequest(
+        @Size(min = 1, max = 50) String name,
+        @Min(1000) @Max(9999) Integer studentNumber
+) { }
