+++ src/main/java/com/pick/zick/domain/user/dto/responce/UserResponse.java
@@ -1,0 +1,8 @@
+package com.pick.zick.domain.user.dto.responce;
+
+public record UserResponse(
+        String userId,
+        String name,
+        Integer studentNumber
+) {
+}
