--- src/main/java/com/pick/zick/exception/BusinessException.java
@@ -1,13 +1,0 @@
-package com.pick.zick.exception;
-
-import lombok.Getter;
-
-@Getter
-public class BusinessException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-}
