+++ src/main/java/com/pick/zick/global/config/JpaConfig.java
@@ -1,0 +1,4 @@
+package com.pick.zick.global.config;
+
+public class JpaConfig {
+}
