+++ src/main/java/com/pick/zick/global/exception/BusinessException.java
@@ -1,0 +1,13 @@
+package com.pick.zick.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
