+++ src/main/java/com/pick/zick/global/exception/GlobalExceptionHandler.java
@@ -1,0 +1,20 @@
+package com.pick.zick.global.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import java.util.HashMap;
+import java.util.Map;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(BusinessException.class)
+    public ResponseEntity<Map<String, Object>> handleBusinessException(BusinessException ex) {
+        Map<String, Object> body = new HashMap<>();
+        body.put("type", ex.getErrorCode().name());
+        body.put("statusCode", ex.getErrorCode().getStatus());
+        body.put("message", ex.getErrorCode().getMessage());
+        return ResponseEntity.status(ex.getErrorCode().getStatus()).body(body);
+    }
+}
