+++ src/main/java/com/pick/zick/global/security/userdetails/SecurityUserDetailsService.java
@@ -1,0 +1,20 @@
+package com.pick.zick.global.security.userdetails;
+
+import com.pick.zick.domain.user.persistence.entity.User;
+import com.pick.zick.domain.user.persistence.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.*;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class SecurityUserDetailsService implements UserDetailsService {
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User u = userRepository.findById(username)
+                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다: " + username));
+        return SecurityUserDetails.from(u);
+    }
+}
