--- src/main/java/com/pick/zick/peristence/response/SignupResponse.java
@@ -1,8 +1,0 @@
-package com.pick.zick.peristence.response;
-
-public record SignupResponse(String userId, String message) {
-    public static SignupResponse ok(String id){
-        return new SignupResponse(id, "회원가입이 완료되었습니다.");
-    }
-}
-
