--- src/main/java/com/pick/zick/peristence/response/UserResponse.java
@@ -1,8 +1,0 @@
-package com.pick.zick.peristence.response;
-
-public record UserResponse(
-        String userId,
-        String name,
-        Integer studentNumber
-) {
-}
