--- src/main/java/com/pick/zick/presentation/controller/AuthController.java
@@ -1,22 +1,0 @@
-package com.pick.zick.presentation.controller;
-
-import com.pick.zick.peristence.request.LoginRequest;
-import com.pick.zick.peristence.response.LoginResponse;
-import com.pick.zick.service.AuthService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/auth")
-public class AuthController {
-
-    private final AuthService authService;
-
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponse> login(@RequestBody @Valid LoginRequest req) {
-        return ResponseEntity.ok(authService.login(req));
-    }
-}
