--- src/main/java/com/pick/zick/presentation/controller/UserController.java
@@ -1,23 +1,0 @@
-package com.pick.zick.presentation.controller;
-
-import com.pick.zick.domain.entity.User;
-import com.pick.zick.domain.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/auth")
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserRepository userRepository;
-
-    @PostMapping("/signup")
-    public ResponseEntity<User> signup(@RequestBody User user) {
-        // 단순 저장 (테스트용)
-        User saved = userRepository.save(user);
-        return ResponseEntity.ok(saved);
-    }
-}
-
